initial commit for application agent.
It works acceptable so far in the notebook.

TODOS :
directory structure for unique jobs being searching and used to apply for jobs.
    setup a directory structure for each unique company applied to
    setup a directory structure for each unique job applied to with companyname as directory and JMangabat_jobtitle_date as filename.

setting the notebook up as a OOP structure, so that the search query is the only thing needed generate an application.
setting op scrapers for different job boards; indeed, glassdoor, graduateland, etc.
    setup dockerdb for the job board data.